+++ src/core/api/dns_api.cpp
@@ -35,7 +35,8 @@
 
 #include <openthread/dns_client.h>
 
-#include "common/as_core_type.hpp"
+#include "common/instance.hpp"
+#include "net/dns_types.hpp"
 
 using namespace ot;
