+++ src/core/api/link_metrics_api.cpp
@@ -38,6 +38,7 @@
 #include <openthread/link_metrics.h>
 
 #include "common/as_core_type.hpp"
+#include "common/locator_getters.hpp"
 
 using namespace ot;
