+++ src/core/utils/srp_client_buffers.hpp
@@ -36,8 +36,11 @@
 
 #include "openthread-core-config.h"
 
+#if OPENTHREAD_CONFIG_SRP_CLIENT_BUFFERS_ENABLE
+
 #include <openthread/srp_client_buffers.h>
 
+#include "common/as_core_type.hpp"
 #include "common/clearable.hpp"
 #include "common/locator.hpp"
 #include "common/non_copyable.hpp"
@@ -46,8 +49,6 @@
 
 namespace ot {
 namespace Utils {
-
-#if OPENTHREAD_CONFIG_SRP_CLIENT_BUFFERS_ENABLE
 
 #if !OPENTHREAD_CONFIG_SRP_CLIENT_ENABLE
 #error "OPENTHREAD_CONFIG_SRP_CLIENT_BUFFERS_ENABLE requires OPENTHREAD_CONFIG_SRP_CLIENT_ENABLE feature."
@@ -260,9 +261,12 @@ private:
     Pool<ServiceEntry, kMaxServices> mServicePool;
 };
 
-#endif // OPENTHREAD_CONFIG_SRP_CLIENT_BUFFERS_ENABLE
-
 } // namespace Utils
+
+DefineCoreType(otSrpClientBuffersServiceEntry, Utils::SrpClientBuffers::ServiceEntry);
+
 } // namespace ot
+
+#endif // OPENTHREAD_CONFIG_SRP_CLIENT_BUFFERS_ENABLE
 
 #endif // SRP_CLIENT_BUFFERS_HPP_
